requirements.txt
Django==4.2
djangorestframework==3.14.0
Markdown==3.5.2
django-filter==24.1
psycopg2==2.9.9
django-environ==0.11.2
